Change the method of receiving data to accomodate downloads better
This fixes a problem where a chunk of received data was assume to be always binary or always messages. This wasn't always the case, so created a new method for receive_data_chunk which is overridden for downloads.

The method needs_parsing? then dictates whether the data should be parsed or just sent directly to this method. This is mostly true except for when downloading where the data need not be parsed.

Correctly downloaded a file which previously failed