Add a websocket subscription service.
I've found that subscribing to the channel via DRb is extremely buggy and was a pain in the ass to get around. This way seems pretty simple and fairly generalized. Any client can subscribe to the stream and it's supported by default in the em-http-request library.

Makes sense to me and this will most likely percolate over to dcfs soon when I get around to it...